Use LONG instead of DWORD for semaphore and wait counts
As the semaphore count returned by CSemaphoreState::CAvail() and
m_cAvail have originally been signed, let's not unnecessarily switch
to an unsigned type.  Doing so also allows us to remove some type casts.

+++ dev/ese/published/inc/sync.hxx
@@ -1997,9 +1997,9 @@ class CSemaphoreState
 
         //    accessors
 
-        const INT CAvail() const { return (INT)m_cAvail; }
-        const INT CWait() const { return (INT)m_cWait; }
-        volatile DWORD * GetAvailAddress() { return &m_cAvail; }
+        const INT CAvail() const { return m_cAvail; }
+        const INT CWait() const { return m_cWait; }
+        volatile LONG * GetAvailAddress() { return &m_cAvail; }
 
         //    debugging support
 
@@ -2017,8 +2017,8 @@ class CSemaphoreState
 
             struct
             {
-                volatile DWORD  m_cAvail;
-                volatile DWORD  m_cWait;
+                volatile LONG   m_cAvail;
+                volatile LONG   m_cWait;
             };
         };
 };
@@ -2137,8 +2137,8 @@ class CSemaphore
         const BOOL _FTryAcquire( const INT cSpin );
         const BOOL _FAcquire( const DWORD dwTimeout );
         void _Release( const INT cToRelease );
-        const BOOL _FWait( const DWORD cAvail, const DWORD dwTimeout );
-        const BOOL _FOSWait( const DWORD cAvail, const DWORD dwTimeout );
+        const BOOL _FWait( const LONG cAvail, const DWORD dwTimeout );
+        const BOOL _FOSWait( const LONG cAvail, const DWORD dwTimeout );
 };
 
 //  acquire one count of the semaphore, waiting forever if necessary
